+++ typus/models/__init__.py
@@ -1,9 +1,13 @@
 from .detection import ImageDetectionResult, InstancePrediction
-from .geometry import BBox, EncodedMask
+from .geometry import BBox, EncodedMask, BBoxXYWHNorm, BBoxMapper, to_xyxy_px, from_xyxy_px
 
 __all__ = [
     "BBox",
     "EncodedMask",
     "InstancePrediction",
     "ImageDetectionResult",
+    "BBoxXYWHNorm",
+    "BBoxMapper", 
+    "to_xyxy_px",
+    "from_xyxy_px",
 ]